Updated user model. Refactoring admin dashboard display. Added new logo.
