+++ client/src/pages/adminDashboard/userSettings/UserSettingsTable.js
@@ -1,0 +1,286 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {useNavigate} from "react-router-dom";
+import {Badge, Button, Dropdown} from "react-bootstrap";
+import React from "react";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
+import {
+  faCheckCircle,
+  faEdit,
+  faInfoCircle,
+  faRedo,
+  faTimesCircle
+} from "@fortawesome/free-solid-svg-icons";
+import ToolkitProvider, {Search} from "react-bootstrap-table2-toolkit";
+import BootstrapTable from "react-bootstrap-table-next";
+import paginationFactory from "react-bootstrap-table2-paginator";
+import PropTypes from "prop-types";
+import swal from "sweetalert";
+import axios from "axios";
+
+const UserSettingsTable = ({users, showModal}) => {
+
+  const navigate = useNavigate();
+
+  const resetUserPassword = (user) => {
+    swal({
+      title: "Are you sure you want to reset the password for user: "
+          + user["displayName"] + " (" + user["email"] + ")?",
+      text: "This will override any existing password reset requests and send a "
+          + "new notification email to the user.",
+      icon: "warning",
+      buttons: true
+    })
+    .then(val => {
+      if (val) {
+        axios.post("/api/internal/user/" + user["id"] + "/password-reset")
+        .then(() => {
+          swal("Password reset successful",
+              "A notification email has been sent to the user.",
+              "success")
+        })
+        .catch(error => {
+          console.error(error);
+          swal("Request failed",
+              "Check the server log for more information.",
+              "warning");
+        })
+      }
+    })
+  }
+
+  const toggleUserActive = (user, active) => {
+    swal({
+      title: "Are you sure you want to " + (!!active ? "enable" : "disable")
+          + " user: " + user["displayName"] + " (" + user["email"] + ")?",
+      text: "Disabled users cannot be added to new studies and assays, but they "
+          + "will remain associated with existing studies and assays.",
+      icon: "warning",
+      buttons: true
+    })
+    .then(val => {
+      if (val) {
+        axios.post("/api/internal/user/" + user["id"] + "/status?active=" + active)
+        .then(() => {
+          swal("User " + (!!active ? "enabled" : "disabled"),
+              "Refresh the page to view the updated user information.",
+              "success")
+        })
+        .catch(error => {
+          console.error(error);
+          swal("Request failed",
+              "Check the server log for more information.",
+              "warning");
+        })
+      }
+    });
+  }
+
+  const columns = [
+    {
+      dataField: "displayName",
+      text: "Display Name",
+      sort: true,
+      headerStyle: {width: '30%'},
+      formatter: (cell, d) => d.displayName,
+      sortFunc: (a, b, order, dataField, rowA, rowB) => {
+        if (rowA.displayName > rowB.displayName) {
+          return order === "desc" ? -1 : 1;
+        }
+        if (rowB.displayName > rowA.displayName) {
+          return order === "desc" ? 1 : -1;
+        }
+        return 0;
+      },
+    },
+    {
+      dataField: "email",
+      text: "Email",
+      sort: true,
+      headerStyle: {width: '30%'},
+      formatter: (c, d) => <Button variant={"link"}
+                                         onClick={() => showModal(d)}>{d.email}</Button>,
+      sortFunc: (a, b, order, dataField, rowA, rowB) => {
+        if (rowA.email > rowB.email) {
+          return order === "desc" ? -1 : 1;
+        }
+        if (rowB.email > rowA.email) {
+          return order === "desc" ? 1 : -1;
+        }
+        return 0;
+      },
+    },
+    {
+      dataField: "type",
+      text: "Type",
+      sort: true,
+      headerStyle: {width: '10%'},
+      formatter: (c, d) => {
+        if (d.type === "STANDARD_USER") {
+          return <Badge bg="success">Standard</Badge>
+        } else if (d.type === "API_USER") {
+          return <Badge bg="warning">API</Badge>
+        } else {
+          return <Badge bg="danger">System</Badge>
+        }
+      }
+    },
+    {
+      dataField: "admin",
+      text: "Role",
+      sort: true,
+      headerStyle: {width: '10%'},
+      formatter: (c, d) => {
+        if (d.admin) {
+          return <Badge bg="danger">Admin</Badge>
+        } else {
+          return <Badge bg="info">User</Badge>
+        }
+      }
+    },
+    {
+      dataField: "status",
+      text: "Status",
+      sort: true,
+      headerStyle: {width: '10%'},
+      formatter: (c, d) => {
+        if (d.locked) {
+          return <Badge bg="warning">Locked</Badge>
+        } else if (d.active) {
+          return <Badge bg="success">Active</Badge>
+        } else {
+          return <Badge bg="danger">Inactive</Badge>
+        }
+      }
+    },
+    {
+      dataField: "controls",
+      text: "",
+      sort: false,
+      headerStyle: {width: '10%'},
+      formatter: (c, d) => {
+        return (
+            <React.Fragment>
+
+              <Dropdown>
+
+                <Dropdown.Toggle variant={"outline-primary"}>
+                  {/*<FontAwesomeIcon icon={faBars} />*/}
+                  &nbsp;Options&nbsp;
+                </Dropdown.Toggle>
+
+                <Dropdown.Menu>
+
+                  <Dropdown.Item onClick={() => showModal(d)}>
+                    <FontAwesomeIcon icon={faInfoCircle}/>
+                    &nbsp;&nbsp;
+                    View Details
+                  </Dropdown.Item>
+
+                  <Dropdown.Item
+                      onClick={() => navigate("/users/" + d.id + "/edit")}
+                  >
+                    <FontAwesomeIcon icon={faEdit}/>
+                    &nbsp;&nbsp;
+                    Edit User
+                  </Dropdown.Item>
+
+                  <Dropdown.Divider/>
+
+                  {
+                    !!d.active ? (
+                        <Dropdown.Item
+                            className={"text-warning"}
+                            onClick={() => toggleUserActive(d, false)}
+                        >
+                          <FontAwesomeIcon icon={faTimesCircle}/>
+                          &nbsp;&nbsp;
+                          Set Inactive
+                        </Dropdown.Item>
+                    ) : (
+                        <Dropdown.Item
+                            className={"text-warning"}
+                            onClick={() => toggleUserActive(d, true)}
+                        >
+                          <FontAwesomeIcon icon={faCheckCircle}/>
+                          &nbsp;&nbsp;
+                          Set Active
+                        </Dropdown.Item>
+                    )
+                  }
+
+                  <Dropdown.Item
+                      className={"text-warning"}
+                      onClick={() => resetUserPassword(d)}
+                  >
+                    <FontAwesomeIcon icon={faRedo}/>
+                    &nbsp;&nbsp;
+                    Reset Password
+                  </Dropdown.Item>
+
+                </Dropdown.Menu>
+              </Dropdown>
+
+            </React.Fragment>
+        )
+      }
+    }
+  ];
+
+  return (
+      <ToolkitProvider
+          keyField="id"
+          data={users}
+          columns={columns}
+          search
+          exportCSV
+      >
+        {props => (
+            <div>
+              <div className="float-end">
+                <Search.SearchBar
+                    {...props.searchProps}
+                />
+              </div>
+              <BootstrapTable
+                  bootstrap4
+                  keyField="id"
+                  bordered={false}
+                  pagination={paginationFactory({
+                    sizePerPage: 10,
+                    sizePerPageList: [10, 20, 40, 80]
+                  })}
+                  defaultSorted={[{
+                    dataField: "displayName",
+                    order: "asc"
+                  }]}
+                  {...props.baseProps}
+              >
+              </BootstrapTable>
+            </div>
+        )}
+      </ToolkitProvider>
+  )
+
+};
+
+UserSettingsTable.propTypes = {
+  users: PropTypes.array.isRequired,
+  showModal: PropTypes.func.isRequired,
+}
+
+export default UserSettingsTable;
