+++ web/src/main/java/io/studytracker/mapstruct/dto/api/UserPayloadDto.java
@@ -16,6 +16,7 @@
 
 package io.studytracker.mapstruct.dto.api;
 
+import io.studytracker.model.UserType;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
@@ -30,6 +31,7 @@ public class UserPayloadDto {
   private @NotBlank String email;
   private String department;
   private String title;
+  private UserType type;
   private boolean admin = false;
   private Date createdAt;
   private Date updatedAt;
